Reverted to edit df78a79c-7bdf-4030-aa17-a652943652a6: "Fix HTML popout window issue
Updated the CodePlayground component to ensure that the popout window functionality works correctly with HTML code execution, in addition to JavaScript.
[skip gpt_engineer]"
